feat: feito detalhar de produto
