+++ package.json
@@ -8,6 +8,8 @@
     "@radix-ui/react-select": "^2.0.0",
     "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slot": "^1.0.2",
+    "@tanstack/react-query": "^5.22.2",
+    "@tanstack/react-query-devtools": "^5.24.0",
     "axios": "^1.5.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
@@ -26,6 +28,7 @@
     "tailwindcss-animate": "^1.0.7",
     "validator": "^13.11.0",
     "web-vitals": "^2.1.4",
+    "zod": "^3.22.4",
     "zustand": "^4.5.1"
   },
   "scripts": {
