+++ src/App.css
@@ -31,10 +31,8 @@ section{
 
 @layer base {
   :root { 
-/*     --background: 0 0% 100%;
+    --background: 0 0% 100%;
     --foreground: 222.2 47.4% 11.2%;
-    --background: 224 71% 4%;
-    --foreground: 213 31% 91%;
 
     --text: 0, 0%, 100%;
  
@@ -65,8 +63,8 @@ section{
  
     --ring: 215 20.2% 65.1%;
  
-    --radius: 0.5rem; */
-    --background: 224 71% 4%;
+    --radius: 0.5rem;
+   /*  --background: 224 71% 4%;
     --foreground: 213 31% 91%;
  
     --muted: 223 47% 11%;
@@ -95,7 +93,7 @@ section{
  
     --ring: 216 34% 17%;
  
-    --radius: 0.5rem;
+    --radius: 0.5rem; */
   }
  
   .dark {
