+++ src/components/basicosComponents/Forms/index.jsx
@@ -1,456 +1,330 @@
-import React from 'react'
-import './forms.scss';
-import { ButtonLogin, ButtonCadastrar, ButtonForm, ButtonRemove } from '../../basicosComponents/Buttons/';
-import { ButtonAdd } from '../../basicosComponents/Buttons/';
+import React, { useState } from "react";
+import "./forms.scss";
+import {
+  ButtonLogin,
+  ButtonCadastrar,
+  ButtonForm,
+  ButtonFakeLoading,
+} from "../Buttons";
 
-import { ImHome } from 'react-icons/im';
-import { FaLocationDot } from 'react-icons/fa6';
-import { BsFillTelephoneInboundFill } from 'react-icons/bs';
-import { AiFillClockCircle } from 'react-icons/ai';
+import { ImHome } from "react-icons/im";
 
-import axios from "axios"
-import { useForm } from 'react-hook-form'
+import axios from "axios";
+import { useForm } from "react-hook-form";
 import { useQuery, useMutation } from "react-query";
-import validator from 'validator';
-import { urlApi } from '../../../constants/urlApi';
-const est_id = localStorage.getItem('est_id');
-const url = urlApi
+import validator from "validator";
+import { urlApi } from "../../../constants/urlApi";
+import { DivInput } from "../DivInput";
+import { Label } from "../Label";
+import LoaderSpin from "../loaders/loader-spin";
+import Toast from "../toast";
+import { Card, CardContent, CardHeader, CardTitle } from "../../../componentes/card";
+import { Input } from "../../../componentes/input";
+import { TextField } from "../../../componentes/text-field";
+import { Form } from "../../../componentes/form";
+const est_id = localStorage.getItem("est_id");
+const url = urlApi;
 
 const FormEstabelecimento = () => {
- /*  const { estIdStore } = useEstIdStore(); */
-  const { register, handleSubmit, reset, getValues, formState: { errors } } = useForm();
-  //pegar dados ja registrados para mostrar
-  const { data, isLoading, refetch } = useQuery(["estabelecimento", est_id], () => {
-    return axios.get(`${url}api/estabelecimentos/geral/${est_id}`, {
-      headers: {
-        'token': localStorage.getItem('token'),
-      },
-    })
-      .then((response) => response.data);
-  })
+  /*  const { estIdStore } = useEstIdStore(); */
+  const [imagemSelecionada, setImagemSelecionada] = useState(null);
+  const [logoSelecionada, setLogoSelecionada] = useState(null);
+
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm();
+
+  // Pegar dados já registrados para mostrar
+  const { data, isLoading } = useQuery(["estabelecimento", est_id], () => {
+    return axios
+      .get(`${url}api/estabelecimentos/geral/${est_id}`, {
+        headers: {
+          token: localStorage.getItem("token"),
+        },
+      })
+      .then((response) => {
+        return response.data;
+      });
+  });
+
+  const [imageSetFromData, setImageSetFromData] = useState(false);
+  if (data && data.logo && !imageSetFromData) {
+    setImagemSelecionada(data.logo);
+    setImageSetFromData(true);
+  }
 
   const onSubmit = (dataForm) => {
     const formData = new FormData();
-    formData.append('nome', dataForm.nome);
-    formData.append('descricao', dataForm.descricao);
-    formData.append('logo', dataForm.logo[0]);
+    formData.append("nome", dataForm.nome);
+    formData.append("descricao", dataForm.descricao);
+    formData.append("logo", logoSelecionada);
+    // @ts-ignore
     mutate(formData);
-  }
-  const { mutate } = useMutation((formData) => {
-    return axios.put(`${url}api/estabelecimentos/geral/${est_id}`, formData, {
-      headers: {
-        'Content-Type': 'multipart/form-data',
-        'token': localStorage.getItem('token'),
-      },
-    })
+  };
+
+  const { mutate, isSuccess, isError } = useMutation((formData) => {
+    return axios
+      .put(`${url}api/estabelecimentos/geral/${est_id}`, formData, {
+        headers: {
+          "Content-Type": "multipart/form-data",
+          token: localStorage.getItem("token"),
+        },
+      })
       .then((response) => response.data);
-  },
-    {
-      onSuccess: (responseData) => {
-        window.alert("sucesso")
-        const dados = responseData;
-      }
+  }, {});
+
+  //pega a imagem selecionada
+  const handleImagemChange = (e) => {
+    const arquivoSelecionado = e.target.files[0];
+    setLogoSelecionada(e.target.files[0]);
+    if (arquivoSelecionado) {
+      const reader = new FileReader();
+
+      reader.onload = (e) => {
+        setImagemSelecionada(e.target.result);
+      };
+
+      reader.readAsDataURL(arquivoSelecionado);
     }
-  );
+  };
+  //retira a imagem selecionada
+  const handleRemoverImagem = () => {
+    setImagemSelecionada(null);
+  };
+
+  if (isLoading) {
+    return <LoaderSpin />;
+  }
 
   return (
-    <div className="formContainer">
-      <div className="titulo">
-        <ImHome />
-        <h3>Informações básicas</h3>
-      </div>
+    <Card 
+    /* w-full lg:w-[60%]  */
+    className="
+    mt-20 lg:mt-0">
+      <CardHeader>
+        <CardTitle  className="flex justify-center gap-2">
+          <ImHome />
+          Informações Básicas
+        </CardTitle>
+      </CardHeader>
 
-      <form onSubmit={handleSubmit(onSubmit)} className="form">
-        <div className="input">
-          <label htmlFor="nome">Nome do estabelecimento</label>
-          <input
-            type="text"
-            name="nome"
-            id="nome"
-            placeholder="Lanchonete"
-            className={errors?.nome && "inputError"}
-            defaultValue={data?.nome || ""}
-            {...register("nome", { required: true })}
-          />
-          {errors?.nome?.type === 'required' && <p className='errorMessage'>Nome é requirido</p>}
-        </div>
+      <CardContent>
+        <form
+          onSubmit={handleSubmit(onSubmit)}
+          className="flex flex-col items-center justify-center gap-4"
+        >
+          <DivInput>
+            <Label htmlFor="nome">Nome do estabelecimento</Label>
+            <Input
+              type="text"
+              name="nome"
+              id="nome"
+              placeholder="Lanchonete"
+              className={` ${
+                errors?.nome && "outline-rose-500"
+              }`}
+              defaultValue={data.nome}
+              {...register("nome", { required: true })}
+            />
+          </DivInput>
+          <DivInput>
+            <Label htmlFor="descricao">Descrição do estabelecimento</Label>
+            <textarea
+              name="descricao"
+              id="descricao"
+              cols="50"
+              rows="3"
+              placeholder="Mostre o que seu estabelecimento tem de melhor aos seus clientes"
+              className={`rounded-md p-2 w-full outline-primary border border-input "
+              ${
+                errors?.nome && "outline-rose-500"
+              }`}
+              defaultValue={data?.descricao}
+              {...register("descricao")}
+            ></textarea>
+            {errors?.nome?.type === "required" && (
+              <p className="text-rose-500">Nome é requerido</p>
+            )}
+          </DivInput>
 
-        <div className="input">
-          <label htmlFor="descricao">Descrição do estabelecimento</label>
-          <textarea
-            name="descricao"
-            id="descricao"
-            cols="50"
-            rows="3"
-            placeholder="Mostre o que seu estabelecimento tem de melhor aos seus clientes"
-            defaultValue={data?.descricao || ""}
-            {...register("descricao")}
-          ></textarea>
-        </div>
+          <div>
+            {imagemSelecionada ? (
+              <div>
+                <img
+                  src={imagemSelecionada}
+                  alt="Imagem Selecionada"
+                  className="w-40 h-36 mb-1 bg-background object-cover"
+                />
+              </div>
+            ) : (
+              <div>
+                <label
+                  className="w-40 h-36 bg-background p-3 rounded inline-block cursor-pointer mb-1"
+                  htmlFor="customFile"
+                >
+                  Clique Aqui para Adicionar uma imagem
+                </label>
+              </div>
+            )}
 
-        <div className="custom-file">
-          <label className="custom-file-label" htmlFor="customFile">
-            Escolha um logotipo
-          </label>
-          <input type="file" className="custom-file-input"
-            id="customFile"
-            name="file"
-            {...register("logo")}
-          />
-        </div>
+            {imagemSelecionada ? (
+              <label
+                className="bg-accent p-1 rounded inline-block cursor-pointer z-50"
+                onClick={handleRemoverImagem}
+              >
+                Remover
+              </label>
+            ) : (
+              <>
+                <label
+                  className="bg-background p-1 rounded inline-block cursor-pointer"
+                  htmlFor="customFile"
+                >
+                  Adicionar
+                </label>
+                <input
+                  type="file"
+                  className="absolute opacity-0 cursor-pointer"
+                  id="customFile"
+                  name="file"
+                  {...register("logo")}
+                  onChange={handleImagemChange}
+                />
+              </>
+            )}
+          </div>
 
-        <ButtonForm />
-      </form>
-    </div>
+          <ButtonForm />
+          </form>
+        </CardContent>
+
+
+        {isSuccess && (
+          <Toast type="success" duration={4000}>
+            Informações Básicas alteradas com Sucesso!
+          </Toast>
+        )}
+    </Card>
   );
 };
 
-const FormLocalizacao = () => {
-  const { register, handleSubmit, reset, getValues } = useForm();
-  const onSubmit = (dataLocalizacao) => {
-    mutate(dataLocalizacao);
-  }
-
-  //pegar dados ja registrados para mostrar
-  const { data, isLoading, refetch } = useQuery(["localizacao", est_id], () => {
-    return axios.get(`${url}api/localizacao/${est_id}`, {
-      headers: {
-        'token': localStorage.getItem('token'),
-      },
-    })
-      .then((response) => response.data);
-  })
-
-  const { mutate } = useMutation((dataLocalizacao) => {
-    return axios.put(`${url}api/localizacao/${est_id}`, dataLocalizacao, {
-      headers: {
-        'token': localStorage.getItem('token'),
-      },
-    })
-      .then((response) => response.data);
-  },
-    {
-      onSuccess: (responseData) => {
-        const dados = responseData;
-      }
-    }
-  );
+const FormCadastro = ({
+  handleSubmit,
+  onSubmit,
+  register,
+  errors,
+  isLoading,
+}) => {
   return (
-    <div className="formContainer">
-      <div className="titulo">
-        <FaLocationDot />
-        <h3>Localização do estabelecimento</h3>
-      </div>
+    <form
+      onSubmit={handleSubmit(onSubmit)}
+      className="flex flex-col items-center justify-center gap-4 w-80 sm:w-96"
+    >
+      <DivInput>
+        <Label htmlFor="nome">Seu nome</Label>
+        <Input
+          type="text"
+          name="nome"
+          id="nome"
+          className={`rounded p-2 w-full ${errors?.nome && "outline-rose-500"}`}
+          {...register("nome", { required: true })}
+        />
+        {errors?.nome?.type === "required" && (
+          <p className="text-rose-500">Nome é requerido</p>
+        )}
+      </DivInput>
 
-      <form onSubmit={handleSubmit(onSubmit)} className="form">
+      <DivInput>
+        <Label htmlFor="email">E-mail</Label>
+        <Input
+          type="text"
+          name="email"
+          id="email"
+          className={`rounded-md p-2 w-full  ${
+            errors?.email && "outline-rose-500"
+          }`}
+          {...register("email", {
+            required: true,
+            validate: (value) => validator.isEmail(value),
+          })}
+        />
+        {errors?.email?.type === "required" && (
+          <p className="text-rose-500">E-mail é requerido</p>
+        )}
+        {errors?.email?.type === "validate" && (
+          <p className="ext-rose-500">E-mail Inválido</p>
+        )}
+      </DivInput>
 
-        <div className="input">
-          <label htmlFor="cep">CEP</label>
-          <input type="text" name="cep" id="cep"
-            {...register("cep")}
-            defaultValue={data?.cep || ""}
-          />
-        </div>
-
-        <div className="input">
-          <label htmlFor="endereco">Endereço</label>
-          <input type="text" name="endereco" id="endereco"
-            defaultValue={data?.endereco || ""}
-            {...register("endereco")}
-          />
-        </div>
-
-        <div className="input">
-          <label htmlFor="numero">Número</label>
-          <input type="number" name="numero" id="numero"
-            {...register("numero")}
-            defaultValue={data?.numero || ""}
-          />
-        </div>
-
-        <div className="input">
-          <label htmlFor="bairro">Bairro</label>
-          <input type="text" name="bairro" id="bairro"
-            {...register("bairro")}
-            defaultValue={data?.bairro || ""}
-          />
-        </div>
-
-        <div className="input">
-          <label htmlFor="cidade">Cidade</label>
-          <input type="text" name="cidade" id="cidade"
-            {...register("cidade")}
-            defaultValue={data?.cidade || ""}
-          />
-        </div>
-        <ButtonForm />
-      </form>
-    </div>
-  )
-}
-
-const FormContato = ({ setPassoAtual, handleVoltar }) => {
-  const { register, handleSubmit, reset, getValues } = useForm();
-  const onSubmit = (data) => {
-    mutate(data);
-  }
-
-  //pegar contatos ja registrados para mostrar
-  const { data, isLoading, refetch } = useQuery(["contatos", est_id], () => {
-    return axios.get(`${url}api/contatos/${est_id}`, {
-      headers: {
-        'token': localStorage.getItem('token'),
-      },
-    })
-      .then((response) => response.data);
-  })
-
-  const { mutate: deleteContato } = useMutation(
-    (contatoId) =>
-      axios.delete(`${url}contatos/${contatoId}`, {
-        headers: {
-          'token': localStorage.getItem('token'),
-        },
-      }),
-    {
-      onSuccess: () => {
-        refetch(); 
-      },
-      onError: (error) => {
-        console.error('Erro ao excluir o contato', error);
-      },
-    }
+      {!isLoading ? <ButtonCadastrar /> : <ButtonFakeLoading />}
+    </form>
   );
+};
 
-  const { mutate } = useMutation((data) => {
-    return axios.post(`${url}contatos`, data, {
-      headers: {
-        'token': localStorage.getItem('token'),
-      },
-    })
-      .then((response) => response.data);
-  },
-    {
-      onSuccess: (responseData) => {
-        refetch();//verificar
-        const dados = responseData;
-        reset();
-      }
-    }
+const FormLogin = ({
+  handleSubmit,
+  onSubmit,
+  register,
+  errors,
+  isError,
+  errorLogin,
+}) => {
+  return (
+    <form
+      onSubmit={handleSubmit(onSubmit)}
+      className="flex flex-col items-center justify-center gap-4 w-80 sm:w-96"
+    >
+      {isError && !errorLogin ? (
+        <p>
+          Aconteceu algum erro inesperado ao executar essa ação, se o erro
+          persistir entre em contato com o suporte!"
+        </p>
+      ) : (
+        ""
+      )}
+      {errorLogin && (
+        <p className="text-rose-800 font-medium bg-rose-200 p-2 rounded shadow shadow-rose-200">
+          {errorLogin}
+        </p>
+      )}
+      <DivInput>
+        <Label htmlFor="email">E-mail</Label>
+        <Input
+          type="text"
+          name="email"
+          id="email"
+          {...register("email", {
+            required: true,
+            validate: (value) => validator.isEmail(value),
+          })}
+        />
+        {errors?.email?.type === "required" && (
+          <p className="text-rose-500">E-mail é requirido</p>
+        )}
+        {errors?.email?.type === "validate" && (
+          <p className="text-rose-500">E-mail Invalido</p>
+        )}
+      </DivInput>
+
+      <DivInput className="input">
+        <Label htmlFor="senha">Senha</Label>
+        <Input
+          type="password"
+          name="senha"
+          id="senha"
+          placeholder="********"
+          {...register("senha", { required: true })}
+        />
+
+        {errors?.senha?.type === "required" && (
+          <p className="text-rose-500">senha é requirido</p>
+        )}
+      </DivInput>
+      <ButtonLogin />
+    </form>
   );
-  return (
-    <div className="formContainer">
-      <div className="titulo">
-        <BsFillTelephoneInboundFill />
-        <h3>Meios de contato do estabelecimento</h3>
-      </div>
+};
 
-      <form onSubmit={handleSubmit(onSubmit)} className="form">
-        <div className="inputSelect">
-          <select name="tipoContato" id="tipoContato" {...register("tipo")}>
-            <option value="">Escolha</option>
-            <option value="telefone">Telefone</option>
-            <option value="whatsapp">Whatsapp</option>
-            <option value="email">E-mail</option>
-          </select>
-          <input type="text" name="contato" id="contato" {...register("contato")} />
-          <ButtonAdd />
-        </div>
-      </form>
-      {isLoading === false &&
-        data.map((contato) => (
-          <div className="getContatos" key={contato.id}>
-            <div className="tipo">
-              <p>{contato.tipo}</p>
-            </div>
-
-            <div className="contato">
-              <p>{contato.contato}</p>
-            </div>
-            <ButtonRemove onClick={() => deleteContato(contato.id)}/>
-          </div>
-        ))
-      }
-    </div>
-  )
-}
-
-const FormHorario = () => {
-  const { register, handleSubmit, reset, getValues } = useForm();
-  const onSubmit = (data) => {
-    mutate(data);
-  }
-
-  //pegar horarios ja registrados para mostrar
-  const { data, isLoading, refetch } = useQuery(["horarios", est_id], () => {
-    return axios.get(`${url}api/horarios/${est_id}`, {
-      headers: {
-        'token': localStorage.getItem('token'),
-      },
-    })
-      .then((response) => response.data);
-  })
-
-  const { mutate } = useMutation((data) => {
-    return axios.post(`${url}horarios/`, data, {
-      headers: {
-        'token': localStorage.getItem('token'),
-      },
-    })
-      .then((response) => response.data);
-  },
-    {
-      onSuccess: (responseData) => {
-        const dados = responseData;
-        reset();
-        refetch();//verificar
-      }
-    }
-  );
-
-  const { mutate: deleteHorario } = useMutation(
-    (HorarioId) =>
-      axios.delete(`${url}api/horarios/${HorarioId}`, {
-        headers: {
-          'token': localStorage.getItem('token'),
-        },
-      }),
-    {
-      onSuccess: () => {
-        refetch(); 
-      },
-      onError: (error) => {
-        console.error('Erro ao excluir o contato', error);
-      },
-    }
-  );
-  return (
-    <div className="formContainer">
-      <div className="titulo">
-        <AiFillClockCircle />
-        <h3>Horários de atendimento</h3>
-      </div>
-
-      <form onSubmit={handleSubmit(onSubmit)} className="form">
-        <div className="inputSelect">
-          <select name="dia" id="dia" {...register("dia")}>
-            <option value=""></option>
-            <option value="Sabado">Sabado</option>
-            <option value="Domingo">Domingo</option>
-            <option value="Segunda Feira">Segunda Feira</option>
-            <option value="Terça Feira">Terça Feira</option>
-          </select>
-          <input type="text" name="hor_abre" id="hor_abre" {...register("hor_abre")} />
-          <input type="text" name="hor_fecha" id="hor_fecha" {...register("hor_fecha")} />
-          <ButtonAdd />
-        </div>
-      </form>
-      {isLoading === false &&
-        data.map((horario) => (
-          <div className="getContatos" key={horario.id}>
-            <div className="tipo">
-              <p>{horario.dia}</p>
-            </div>
-
-            <div className="contato">
-              <p>{horario.hor_abre}</p>
-              <p>{horario.hor_fecha}</p>
-            </div>
-
-            <ButtonRemove onClick={() => deleteHorario(horario.id)}/>
-          </div>
-        ))
-      }
-    </div>
-  )
-}
-
-const FormCadastro = ({ handleSubmit, onSubmit, register, errors }) => {
-  return (
-    <div className="formContainer formCadastro">
-      <div className="titulo">
-        <FaLocationDot />
-        <h3>Cadastro</h3>
-      </div>
-
-      <form onSubmit={handleSubmit(onSubmit)} className="form">
-
-        <div className="input">
-          <label htmlFor="nome">Seu nome</label>
-          <input type="text" name="nome"
-            id="nome"
-            className={errors?.nome && "inputError"}
-            {...register("nome", { required: true })}
-          />
-          {errors?.nome?.type === 'required' && <p className='errorMessage'>Nome é requirido</p>}
-        </div>
-
-        <div className="input">
-          <label htmlFor="email">E-mail</label>
-          <input type="text" name="email"
-            id="email"
-            className={errors?.email && "inputError"}
-            {...register("email", {
-              required: true,
-              validate: (value) => validator.isEmail(value)
-            })}
-          />
-          {errors?.email?.type === 'required' && <p className='errorMessage'>E-mail é requirido</p>}
-          {errors?.email?.type === 'validate' && <p className='errorMessage'>E-mail Invalido</p>}
-        </div>
-
-        <ButtonCadastrar />
-      </form>
-    </div>
-  )
-}
-
-const FormLogin = ({ handleSubmit, onSubmit, register, errors }) => {
-  return (
-    <div className="formContainer formCadastro">
-      <div className="titulo">
-        <FaLocationDot />
-        <h3>Login</h3>
-      </div>
-
-      <form onSubmit={handleSubmit(onSubmit)} className="form">
-        <div className="input">
-          <label htmlFor="email">E-mail</label>
-          <input type="text" name="email"
-            id="email"
-            className={errors?.email && "inputError"}
-            {...register("email", {
-              required: true,
-              validate: (value) => validator.isEmail(value)
-            })}
-          />
-          {errors?.email?.type === 'required' && <p className='errorMessage'>E-mail é requirido</p>}
-          {errors?.email?.type === 'validate' && <p className='errorMessage'>E-mail Invalido</p>}
-        </div>
-
-        <div className="input">
-          <label htmlFor="senha">Senha</label>
-          <input
-            type="password"
-            name="senha"
-            id="senha"
-            placeholder="********"
-            className={errors?.senha && "inputError"}
-            {...register("senha", { required: true })}
-          />
-          {errors?.senha?.type === 'required' && <p className='errorMessage'>senha é requirido</p>}
-        </div>
-        <ButtonLogin />
-      </form>
-    </div>
-  )
-}
-
-export {
-  FormEstabelecimento,
-  FormLocalizacao,
-  FormContato,
-  FormHorario,
-  FormCadastro,
-  FormLogin
-}   
+export { FormEstabelecimento, FormCadastro, FormLogin };
