+++ src/services/produto/schemas/doDataProdutoDetalheRequest.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+
+export const DoDataProdutoDetalheSchema = z.object({
+    proId: z.string()
+})
+
+export type DoDataProdutoDetalheRequest = z.infer<typeof DoDataProdutoDetalheSchema>;
